feat: create function to add interest and deduct tax and add to transaction

+++ models/transaction.go
@@ -6,7 +6,7 @@ type Transaction struct {
 	Id                   int       `json:"id" gorm:"primarykey"`
 	SenderWalletNumber   int       `json:"sender_wallet_number"`
 	ReceiverWalletNumber int       `json:"receiver_wallet_number"`
-	Amount               int       `json:"amount"`
+	Amount               float32   `json:"amount"`
 	Description          string    `json:"description"`
 	CreatedAt            time.Time `json:"created_at"`
 }
